move key calc
