+++ config/config.go
@@ -16,6 +16,7 @@ import (
 	"github.com/ava-labs/avalanchego/utils/constants"
 	"github.com/ava-labs/avalanchego/utils/logging"
 	"github.com/ava-labs/avalanchego/utils/set"
+	"github.com/ava-labs/awm-relayer/database"
 	"github.com/ava-labs/awm-relayer/utils"
 
 	"github.com/ava-labs/subnet-evm/ethclient"
@@ -244,25 +245,8 @@ func (c *Config) GetSubnetID(blockchainID ids.ID) ids.ID {
 	return c.blockchainIDToSubnetID[blockchainID]
 }
 
-// TODONOW: Put this somewhere else
-type RelayerKey struct {
-	SourceBlockchainID      ids.ID
-	DestinationBlockchainID ids.ID
-	OriginSenderAddress     common.Address
-	DestinationAddress      common.Address
-}
-
-func (k RelayerKey) CalculateRelayerKey() common.Hash {
-	return utils.CalculateRelayerKey(
-		k.SourceBlockchainID,
-		k.DestinationBlockchainID,
-		k.OriginSenderAddress,
-		k.DestinationAddress,
-	)
-}
-
-func (c *Config) GetAllRelayerKeys() []RelayerKey {
-	var keys []RelayerKey
+func (c *Config) GetAllRelayerKeys() []database.RelayerKey {
+	var keys []database.RelayerKey
 	for _, s := range c.SourceBlockchains {
 		keys = append(keys, s.GetRelayerKeys()...)
 	}
@@ -488,10 +472,10 @@ func (s *SourceBlockchain) GetBlockchainID() ids.ID {
 	return s.blockchainID
 }
 
-func (s *SourceBlockchain) GetRelayerKeys() []RelayerKey {
-	var keys []RelayerKey
+func (s *SourceBlockchain) GetRelayerKeys() []database.RelayerKey {
+	var keys []database.RelayerKey
 	for _, dst := range s.GetSupportedDestinations().List() {
-		keys = append(keys, RelayerKey{
+		keys = append(keys, database.RelayerKey{
 			SourceBlockchainID:      s.GetBlockchainID(),
 			DestinationBlockchainID: dst,
 			OriginSenderAddress:     common.Address{}, // TODO: populate with allowed sender/receiver addresses
